fix: Separate all glosses when going deeper
I first noticed this with 'halkoi', which goes deeper with 'halkoa', but
did not actually differentiate between the noun 'halkoa' and the verb
'halkoa'. I found that odd and decided to take the most naive path
forward (becuase anyone using this dictionary would probably benefit
from knowing that e.g. halkoa could be either a noun or a verb in the
right context).