feat: Change arrows (->) to curly arrows (~>) in going deeper
I've always seen the curly ~> arrow in math to mean "leads to", since
there isn't really a way to represent this otherwise. It's like steps in
an equation: x + 2 = 0 ~> x = -2. I like the way it shows up in my app
with the Fira Code ligatures, and so I decided to change it.