scary: feat: Faster go deeper prefix lookups
Upon code review I realized that we were naively iterating through all
possible go deeper prefixes to find a match, this is obviously silly.
These lookups can be performed in O(1) time with a hashmap. I'm marking
this as "scary" because this is data structure refactoring, though,
which is always apt to cause weird breakages!